pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "clippie"
version = "0.0.1"
dependencies = [
    "numpy",
    "regex",
    "pillow",
]

[project.scripts]
clippie-convert-vocab-file = "clippie.scripts.convert_vocab_file:main"

[tool.setuptools]
include-package-data = true

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
